Significant MorphRange refactor
This commit continues the work I was doing in HTMLBars to create a
single append fast-path for initial render, while still allowing
subsequent passes to run just the dynamic parts.

In HTMLBars, we achieve that by executing a unified set of statements
that include the static bits, and emitting a list of dynamic statements
and a list of the morphs related to them.

Several things were making the initial render slower than it could have
been; for example creating a morph always needed to create an empty
comment node that usually got replaced. This is because morphs always
need to be represented inclusively once they're in the DOM, and APIs
like setNode rely on this characteristic.

This commit creates a second phase in the lifecycle of a morph:
"appending". During the appending phase, a morph is characterized by a
parent and a nextSibling, rather than by a start and end node. During
this phase, only append operations are allowed, and they are always
performed relative to the nextSibling.

This phase needs to be plumbed through into MorphList as well, since it
is important that the initial render of elements of a MorphList can also
participate in this optimization.

Generally speaking, this commit makes state transitions in Morph more
explicit, and operations are generally only allowed on a subset of
statees. For example, while setNode is allowed in the "detached" state,
setRange is not (a range of elements requires a parent, and the
"detached" state doesn't have one).

It also moves towards a more explicit transactional model for changes.

On initial render, consumers of the API need to explicitly enable append
mode by supplying a parent element or parent morph to render into. Once
all of the rendering has completed, they need to explicitly end the
transaction, which will sync first and last nodes correctly across the
tree.

Atomic updates, such as `setNode`, are allowed in the "attached" state
without transactions.

Updates to MorphList after the initial render similarly require a
transactional model (beginMove and finishMove). MorphList will avoid
completely destroying any removed nodes until `finishMove` has
completed, allowing them to be reused reliably.

The intent of this API is for HTMLBars to initialize morphs in append
mode on initial render, finalizing the transaction for each morph once
its associated statement has completed.

MorphList elements should similarly be put into append mode on initial
render, but dynamically created MorphList elements can also be put into
append mode by specifying the morph they will be inserted before. This
allows the unit of work starting from the morphList item to use the same
optimization as initial render.

When re-rendering a regular block, calling `clearForRender` will switch
the block into append mode using the current position in the DOM, and
then clear the contents. This places the morph (and all downstream work)
in the same mode as in initial render. Once the statement has been fully
evaluated, a call to `finishAppend` will commit the transaction.

Many tests are still failing, but I wrote a few tests that help to
demonstrate how the system is intended to work.