Make a path for expecting updates
This allows HTMLBars to say that it is about to render, and if the user
code doesn't trigger any renders, we can clear the contents in
finishAppend.

We don't want to eagerly destroy the node, because if the user happens
to render the same thing, we want to indicate stability and move on.